@@ -8,6 +8,8 @@
     "production": "mix --production"
   },
   "devDependencies": {
+    "@vitejs/plugin-vue": "^5.2.1",
+    "@vue/compiler-sfc": "^3.5.13",
     "axios": "^1.12.2",
     "browser-sync": "^2.29.3",
     "browser-sync-webpack-plugin": "^2.4.0",
@@ -21,16 +23,15 @@
     "resolve-url-loader": "^4.0.0",
     "sass": "^1.93.2",
     "sass-loader": "^16.0.5",
-    "vue": "^2.7.16",
-    "vue-loader": "^15.11.1",
-    "vue-template-compiler": "^2.7.16"
+    "vue": "^3.5.13",
+    "vue-loader": "^17.4.2"
   },
   "dependencies": {
     "@primer/css": "^15.2.0",
     "bourbon": "^7.3.0",
+    "mitt": "^3.0.1",
     "moment": "^2.30.1",
-    "vue-google-adsense": "^1.10.1",
-    "vue-script2": "^2.1.0"
+    "vue-google-adsense": "^2.0.0"
   },
   "packageManager": "pnpm@9.7.1+sha512.faf344af2d6ca65c4c5c8c2224ea77a81a5e8859cbc4e06b1511ddce2f0151512431dd19e6aff31f2c6a8f5f2aced9bd2273e1fed7dd4de1868984059d2c4247"
 }
